fp-in-scala: exercise 3.5 - implement dropWhile; also reverse and filter.
Thank to initially mis-reading the exercise managed to implement filter
with the help of reverse instead.

The important point I missed was that dropWhile operates on the prefix
(or from the head) of the list as long as the predicate is true. I've
mistakenly applied the predicate to the entire list to drop elements
that match the predicate - reverting the dropping of elements to
collecting of elements matching a predicate got me filter instead :)