nps-guide-scala: Actors - default supervisor strategy and actor life-cycle demo.
A Register paper jam is simulated upon which the Barista actor guarding its
child - Register actor - will automatically restart the Register actor as the
default supervisor strategy is OneForOneStrategy - meaning only the single
problematic child actor is restarted.

The see the restarting the actor life-cycle's postRestart is overridden to
output a log message.