nps-guide-scala: Actors - error kernel pattern and mapping error messages.
Its a good idea not to keep state in actors which tasks might fail often - that
is, error prone task should be split into separate (child) actors that don't
require state and can be restarted at will.

In case of an error, a message should be sent back that makes sense for the
actor requesting an action - ie instead the CafeCustomer actor getting back
and AskTimeoutException it will receive a ComeBackLater message if the was
a problem creating a receipt/bill.